(refactor) : GameManager의 playGame 메서드가 GameManagerResponseDto를 반환하도록 변경
