+++ src/main/java/racingcar/controller/RacingCarGameApplication.java
@@ -2,6 +2,7 @@ package racingcar.controller;
 
 import racingcar.domain.GameManager;
 import racingcar.domain.dto.GameManagerRequestDto;
+import racingcar.domain.dto.GameManagerResponseDto;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -9,8 +10,8 @@ public class RacingCarGameApplication {
     public static void main(String[] args) {
         GameManagerRequestDto requestDto = InputView.inputFromUser();
         GameManager gameManager = new GameManager();
-        gameManager.playGame(requestDto);
-        OutputView.printResults(gameManager.getResults());
-        OutputView.printWinners(gameManager.getWinners());
+        GameManagerResponseDto responseDto = gameManager.playGame(requestDto);
+        OutputView.printResults(responseDto.getResults());
+        OutputView.printWinners(responseDto.getWinners());
     }
 }
